+++ Changelog.md
@@ -7,7 +7,8 @@
  - Added recent changes page under the action `recent-changes`. A link can be found in the "More..." menu by default.
  - Changed the cursor when hovering over a time to indicate that the tooltip contains more information.
  - Added icons to the "More..." menu
- - Help section to parsedown parser.
+ - Added help section to parsedown parser.
+ - Added more information to the dev help page.
 
 ## Changed
  - Improved appearance of the all pages list.